// src/web/shared/proto/jaiabot/messages/geographic_coordinate.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.12.0
//   protoc               v3.21.12
// source: jaiabot/messages/geographic_coordinate.proto

/* eslint-disable */

export const protobufPackage = "jaiabot.protobuf";

export interface GeographicCoordinate {
    lat?: number | undefined;
    lon?: number | undefined;
}